+++ docs/api/apiDocs/useSubscribe.tsx
@@ -1,0 +1,9 @@
+/**
+ * @order 10
+ * @name useSubscribe
+ * @pageTitle useSubscribe
+ * @toc false
+ * @groupOrder 10
+ * @embed /defaultContent/_useSubscribe.md
+ */
+export default function App(props: {}) {}