+++ docs/demos/useSubscribe/_basic.tsx
@@ -5,7 +5,6 @@ export default function App() {
   const { render, useSubscribe } = useForm({ name: 'name' });
 
   const renderName = render('name')(<input />);
-
   const subscribeName = useSubscribe(({ model }) => model.name);
 
   return (