+++ docs/types/template/useFormWatch.tsx
@@ -1,7 +1,7 @@
 /**
  * @order 50
- * @embed /embed/_useWatch.md
- * @name UseWatch
+ * @embed /embed/_useFormWatch.md
+ * @name UseFormWatch
  */
 
 export default function UseWatch(props: {}) {}