+++ src/driver/ObserverDriver/abstract/AbstractImp.ts
@@ -1,0 +1,6 @@
+export abstract class AbstractImp {
+  abstract emit(...args: any[]): void;
+  abstract update(...args: any[]): void;
+}
+
+export default AbstractImp;